advanced-react ch24: signup backend
This sets up the app to store the JWTs in a cookie. I thought this was
bad security practice, but it looks like it can be done well:
https://stormpath.com/blog/where-to-store-your-jwts-cookies-vs-html5-web-storage

It looks like the approach here is still vulnerable to CSRF attacks. In
order to harden it, the blog above indicates there are a couple of
things I could do:

1. Use `csurf` (https://github.com/expressjs/csurf) to add and check a
token to requests known to come from this app
2. Check the HTTP `Referer` and `Origin` headers in the API to make sure
they come from an expected domain.