advanced-react: simplify frontend Order logic
The frontend doesn't use the `order.user.id` attribute. It was only
included in the query so that the backend could check permissions.

This is confusing, when consuming the API from the frontend. So instead,
this change allows removing the `user { id }` specification in the
frontend query, and does two separate `order` queries on the backend,
ensuring that the first one can check the permissions, and the second
one returns the attributes needed by the API consumer.

It's somewhat annoying that this requires two async calls rather than
one, but this seemed like the cleanest way to approach the problem.