advanced-react ch62: test RequestReset
Note the `await wait(2)`, which differs from what Wes showed in the
course. His worked with just `await wait()` (equivalent to `await
wait(0)`. I don't understand why mine requires at least a 2ms delay.

And yes, it's consistent with needing these 2 milliseconds. It always
works with 2, and always fails with 1. With 1 or 0, the form has not yet
updated with the success message.