advanced-react ch65: test TakeMyMoney
This requires moving the `handlePayment` function inside the component,
which bums me out. I tried for a while to find a different way to test
this, but I couldn't find one.

It seems like a better way (the best way?) would be to override the
`createOrder` that gets passed in from `Mutation`.