advanced-react: double-fix user.orders bug
In e7c697a13415e09768, I added an `orders { id }` field to frontend's
CURRENT_USER_QUERY. I added this so that it would match the format of
`fakeUser` from `testUtils`.

The problem is that this field didn't exist on the backend. So I added
it there.

But it annoys me when the main code changes for the sake of tests, and
no other tests relied on `fakeUser` having an `orders` attribute. So I
also removed the `orders` key from `CURRENT_USER_QUERY`.